ai-horizon-python/ai-horizon-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-horizon"
version = "0.1.0"
description = "AI-powered classification system for cybersecurity workforce impact analysis"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Habib Jahshan", email = "habib@example.com"}
]
keywords = ["ai", "cybersecurity", "dcwf", "rag", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-genai>=0.8.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pypdf2>=3.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "youtube-transcript-api>=0.6.0",
    "trafilatura>=1.6.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
horizon = "src.main:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
